Conflicts:
	Assignment 2.sdf

+++ Player.h
@@ -44,9 +44,9 @@ public:
 
 	void collectCards(std::vector<Card> &deck); 
 	//empties out the visible cards vector and puts them back in the deck vector
-	int getHandValue(); 
+	double getHandValue(); 
 	// returns the value of the complete 5 card set 
-	int getVisibleHandValue(); 
+	double getVisibleHandValue(); 
 	// evaluates who starts betting by looking at the set of 4 or less cards to find 
 
 	// ui functions
